move misplaced methods
