+++ src/main/java/edu/byu/cs/util/SubmissionUtils.java
@@ -1,14 +1,196 @@
 package edu.byu.cs.util;
 
+import edu.byu.cs.autograder.GradingContext;
+import edu.byu.cs.autograder.GradingException;
+import edu.byu.cs.autograder.git.CommitVerificationConfig;
+import edu.byu.cs.autograder.score.Scorer;
+import edu.byu.cs.dataAccess.DaoService;
 import edu.byu.cs.dataAccess.DataAccessException;
 import edu.byu.cs.dataAccess.ItemNotFoundException;
 import edu.byu.cs.dataAccess.SubmissionDao;
 import edu.byu.cs.model.Phase;
+import edu.byu.cs.model.Rubric;
 import edu.byu.cs.model.Submission;
+import org.eclipse.jgit.annotations.NonNull;
+import org.eclipse.jgit.annotations.Nullable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import java.time.Instant;
 import java.util.HashSet;
 
 public class SubmissionUtils {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(SubmissionUtils.class);
+
+    /**
+     * Approves the highest scoring submissions on the phase so far with a provided penalty percentage.
+     * <br>
+     * This is a simple overload triggering default behavior in the actual method.
+     * @see SubmissionUtils#approveSubmission(String, Phase, String, Integer, Float, Submission).
+     *
+     * @param studentNetId The student to approve
+     * @param phase The phase to approve
+     * @param approverNetId Identifies the TA or professor approving the score
+     * @param penaltyPct The penalty applied for the reduction.
+     *                   This should already be reflected in the `approvedScore` if present.
+     */
+    public static void approveSubmission(
+            @NonNull String studentNetId, @NonNull Phase phase, @NonNull String approverNetId, @NonNull Integer penaltyPct)
+            throws DataAccessException, GradingException {
+        approveSubmission(studentNetId, phase, approverNetId, penaltyPct, null, null);
+    }
+    /**
+     * Approves a submission.
+     * Modifies all existing submissions in the phase with constructed values,
+     * and saves a given value into the grade-book.
+     *
+     * @param studentNetId The student to approve
+     * @param phase The phase to approve
+     * @param approverNetId Identifies the TA or professor approving the score
+     * @param penaltyPct The penalty applied for the reduction.
+     *                   This should already be reflected in the `approvedScore` if present.
+     *                   <p>This will be used to reduce all other submissions,
+     *                   but has no effect in the grade-book when <code>approvedScore</code> is provided.</p>
+     * @param approvedScore <p>The final score that should go in the grade-book.
+     *                      Submissions in the database will <b>not</b> be affected by this value.</p>
+     *                      <p>If `null`, we'll determine this value by applying the penalty to
+     *                      the highest score for any submission in the phase.</p>
+     *                      <p>Provided so that a TA can approve an arbitrary (highest score)
+     *                      submission with a penalty instead of any other fixed rule.</p>
+     * @param submissionToUse Required when `approvedScored` is passed in.
+     *                         Provides a submission which will be used to overwrite the existing score in the grade-book.
+     *                         If a full {@link Submission} object is not available, the {@link Rubric} is only required field in it.
+     */
+    public static void approveSubmission(
+            @NonNull String studentNetId,
+            @NonNull Phase phase,
+            @NonNull String approverNetId,
+            @NonNull Integer penaltyPct,
+            @Nullable Float approvedScore,
+            @Nullable Submission submissionToUse
+    ) throws DataAccessException, GradingException {
+        if (approvedScore != null) {
+            throw new IllegalArgumentException("Passing in non-null approvedScore is not yet implemented.");
+        }
+
+        // Validate params
+        if (studentNetId == null || phase == null || approverNetId == null || penaltyPct == null) {
+            throw new IllegalArgumentException("All of studentNetId, approverNetId, and penaltyPct must not be null.");
+        }
+        if (studentNetId.isBlank() || approverNetId.isBlank()) {
+            throw new IllegalArgumentException("Both studentNetId and approverNetId must not be blank");
+        }
+        if (penaltyPct < 0 || (approvedScore != null && approvedScore < 0)) {
+            throw new IllegalArgumentException("Both penaltyPct and approvedScore must be greater or equal than 0");
+        }
+
+        // Read in data
+        SubmissionDao submissionDao = DaoService.getSubmissionDao();
+        assertSubmissionUnapproved(submissionDao, studentNetId, phase);
+        Submission withheldSubmission = determineSubmissionForConsideration(submissionDao, submissionToUse, studentNetId, phase);
+
+        // Modify values in our database first
+        int submissionsAffected = modifySubmissionEntriesInDatabase(submissionDao, withheldSubmission, approverNetId, penaltyPct);
+
+        // Send score to Grade-book
+        if (approvedScore == null) {
+            approvedScore = SubmissionUtils.prepareModifiedScore(withheldSubmission.score(), penaltyPct);
+        }
+        String gitCommitsComment = "Submission initially blocked due to low commits. Submission approved by admin " + approverNetId;
+        sendScoreToCanvas(withheldSubmission, penaltyPct, gitCommitsComment);
+
+        // Done
+        LOGGER.info("Approved submission for %s on phase %s with score %f. Approval by %s. Affected %d submissions."
+                .formatted(studentNetId, phase.name(), approvedScore, approverNetId, submissionsAffected));
+    }
+
+    private static void assertSubmissionUnapproved(SubmissionDao submissionDao, String studentNetId, Phase phase) throws DataAccessException {
+        Submission withheldSubmission = submissionDao.getFirstPassingSubmission(studentNetId, phase);
+        if (withheldSubmission.isApproved()) {
+            throw new RuntimeException(studentNetId + " needs no approval for phase " + phase);
+        }
+    }
+
+    private static Submission determineSubmissionForConsideration(
+            SubmissionDao submissionDao, Submission providedValue, String studentNetId, Phase phase)
+            throws DataAccessException, GradingException {
+        if (providedValue != null) return providedValue;
+
+        Submission submission = submissionDao.getBestSubmissionForPhase(studentNetId, phase);
+        if (submission == null) {
+            throw new GradingException("No submission was provided nor found for phase " + phase + " with user " + studentNetId);
+        }
+        return submission;
+    }
+
+    private static int modifySubmissionEntriesInDatabase(
+            SubmissionDao submissionDao, Submission withheldSubmission, String approvingNetId, int penaltyPct)
+            throws DataAccessException {
+
+        Float originalScore = withheldSubmission.score();
+        Instant approvedTimestamp = Instant.now();
+        Submission.ScoreVerification scoreVerification = new Submission.ScoreVerification(
+                originalScore,
+                approvingNetId,
+                approvedTimestamp,
+                penaltyPct);
+        int submissionsAffected = SubmissionUtils.approveWithheldSubmissions(
+                submissionDao,
+                withheldSubmission.netId(),
+                withheldSubmission.phase(),
+                scoreVerification);
+
+        if (submissionsAffected < 1) {
+            LOGGER.warn("Approving submissions did not affect any submissions. Something probably went wrong.");
+        }
+        return submissionsAffected;
+    }
+
+    /**
+     * Submits the <code>approvedScore</code> to Canvas by modifying the GIT_COMMITS rubric item.
+     * <br>
+     * In this context, the <code>withheldSubmission</code> isn't <i>strictly</i> necessary,
+     * but it is used to transfer several fields and values that would need to be transferred individually
+     * without it. The most important of these is the <score>field</score>, which will be used to calculate the penalty.
+     *
+     * @param withheldSubmission The baseline submission to approve
+     * @param penaltyPct The percentage that should be reduced from the score for GIT_COMMITS.
+     * @param commitPenaltyComment The comment that should be associated with the GIT_COMMITS rubric, if any.
+     * @throws DataAccessException When the database cannot be accessed.
+     * @throws GradingException When pre-conditions are not met.
+     */
+    private static void sendScoreToCanvas(Submission withheldSubmission, int penaltyPct, String commitPenaltyComment) throws DataAccessException, GradingException {
+        // Prepare and assert arguments
+        if (withheldSubmission == null) {
+            throw new IllegalArgumentException("Withheld submission cannot be null");
+        }
+        Scorer scorer = getScorer(withheldSubmission);
+        scorer.attemptSendToCanvas(withheldSubmission.rubric(), penaltyPct, commitPenaltyComment, true);
+    }
+
+    /**
+     * Constructs an instance of {@link Scorer} using the limited data available in a {@link Submission}.
+     *
+     * @param submission A submission containing context to extract.
+     * @return A constructed Scorer instance.
+     */
+    private static Scorer getScorer(Submission submission) {
+        String studentNetId = submission.netId();
+        Phase phase = submission.phase();
+        if (studentNetId == null) {
+            throw new IllegalArgumentException("Student net ID cannot be null");
+        }
+        if (phase == null) {
+            throw new IllegalArgumentException("Phase cannot be null");
+        }
+
+        return new Scorer(new GradingContext(
+                studentNetId, phase, null, null, null, null,
+                new CommitVerificationConfig(0, 0, 0, 0, 0),
+                null, submission.admin()
+        ));
+    }
 
     /**
      * Updates <b>all</b> of the relevant submissions with an appropriate {@link Submission.ScoreVerification}
